+++ src/index.ts
@@ -38,26 +38,136 @@ export const defaultOptions: BootstrapOptions = {
 	errorHandler: console.error
 };
 
+export class Bootstrap {
+	/**
+	 * Should the process exit when the register/run throw an exception
+	 */
+	public shouldExitOnError = true;
+
+	/**
+	 * Register function runs before entrypoint, and is useful
+	 *	for setting up services etc
+	 */
+	protected register: RunFunction = async () => { };
+
+	/**
+	 * onComplete runs when the run function returns (.then())
+	 */
+	protected onComplete: RunFunction = async () => { };
+
+	/**
+	 * onError runs when the run function throws an exception (.catch())
+	 */
+	protected onError: ErrorHandlerFunction = console.error;
+
+	/**
+	 * onFinally runs when the run function stops (.finally())
+	 */
+	protected onFinally: RunFunction = async () => { };
+
+	/**
+	 * @internal
+	 */
+	protected isBooted = false;
+
+	public constructor(options: BootstrapOptions = defaultOptions) {
+		if (options.register) {
+			this.setRegister(options.register);
+		}
+
+		if (options.onComplete) {
+			this.setOnComplete(options.onComplete);
+		}
+
+		if (options.onFinally) {
+			this.setOnFinally(options.onFinally);
+		}
+
+		if (options.shouldExitOnError !== null) {
+			this.shouldExitOnError = options.shouldExitOnError;
+		}
+
+		if (options.errorHandler) {
+			this.setOnError(options.errorHandler);
+		}
+	}
+
+	/**
+	 * Boot and run a function
+	 *
+	 * @param run Run function
+	 */
+	public boot(run: RunFunction): void {
+		Promise.resolve(this.register ? this.register() : null)
+			.then(async () => await Promise.resolve(run()))
+			.then(async () => await Promise.resolve(this.onComplete()))
+			.catch(async e => {
+				// Handle or log the error
+				await Promise.resolve(this.onError(e));
+
+				if (this.shouldExitOnError) {
+					process.exit(e.code ?? 1);
+				}
+			})
+			.finally(async () => await Promise.resolve(this.onFinally()));
+	}
+
+	/**
+	 * Boot (async)
+	 */
+	public async bootAsync(): Promise<void> {
+		this.shouldExitOnError = false;
+
+		return await new Promise<void>((accept, reject) => {
+			this.setOnError(reject);
+			this.boot(async () => accept());
+		});
+	}
+
+	public setRegister(register: RunFunction): void {
+		if (this.isBooted) {
+			throw new Error('Cannot setRegister() after boot');
+		}
+
+		this.register = register.bind(this);
+	}
+
+	public setOnComplete(onComplete: RunFunction): void {
+		if (this.isBooted) {
+			throw new Error('Cannot setOnComplete() after boot');
+		}
+
+		this.onComplete = onComplete.bind(this);
+	}
+
+	public setOnError(onError: ErrorHandlerFunction): void {
+		if (this.isBooted) {
+			throw new Error('Cannot setOnError() after boot');
+		}
+
+		this.onError = onError.bind(this);
+	}
+
+	public setOnFinally(onFinally: RunFunction): void {
+		if (this.isBooted) {
+			throw new Error('Cannot setOnFinally() after boot');
+		}
+
+		this.onFinally = onFinally.bind(this);
+	}
+}
+
 /**
  * Bootstrap a function
  *
  * @param options Bootstrapping options
  */
-export function bootstrap(options: BootstrapOptions): void {
-	options = { ...defaultOptions, ...options };
+export function bootstrap(options: BootstrapOptions): Bootstrap {
+	const b = new Bootstrap({ ...defaultOptions, ...options });
 
-	Promise.resolve(options.register ? options.register() : null)
-		.then(async () => await Promise.resolve(options.run()))
-		.then(async () => await Promise.resolve(options.onComplete()))
-		.catch(async e => {
-			// Handle or log the error
-			await Promise.resolve(options.errorHandler(e));
+	b.boot(options.run);
 
-			if (options.shouldExitOnError) {
-				process.exit(e.code ?? 1);
-			}
-		})
-		.finally(async () => await Promise.resolve(options.onFinally()));
+	return b;
 }
 
 /**
@@ -70,12 +180,10 @@ export function bootstrap(options: BootstrapOptions): void {
 export async function bootstrapPromise(
 	register: RunFunction
 ): Promise<void> {
-	return await new Promise<void>((accept, reject) => {
-		bootstrap({
-			register: register,
-			run: async () => accept(),
-			errorHandler: async (...e) => reject(...e),
-			shouldExitOnError: false,
-		});
+	const b = new Bootstrap({
+		register: register,
+		run: async () => { },
 	});
+
+	return await b.bootAsync();
 }